Tweaks to the Navbar, login and deletion requests

+++ server/index.js
@@ -9,7 +9,8 @@ const path = require('path');
 const querystring = require('querystring');
 
 const app = express();
-const serverAddress = 'http://ec2-13-57-254-191.us-west-1.compute.amazonaws.com'
+// const serverAddress = 'http://ec2-13-57-254-191.us-west-1.compute.amazonaws.com'
+const serverAddress = 'http://localhost:4000'
 app.use(express.static(path.join(__dirname, '../', 'client', 'dist')))
 app.use(express.static(path.join(__dirname, '../', 'client', 'dist', 'assets')))
 app.use(fileUpload({
@@ -27,30 +28,33 @@ app.listen(4000, () => {
 })
 
 app.get('/api/download/*', (req, res) => {
+  console.log('Recieved download request')
   let file = req.params[0].split('/')
   db.pool.query('SELECT filename FROM files WHERE hash = $1', [file[1]], (err, result) => {
     if (err) {
+      console.log('Download Failed', err)
       res.status(404).send('This file does not exist.')
     } else {
-      res.download(`./db/files/${file[0]}/${result.rows[0].filename}`)
+      let downloadFile = `./db/files/${file[0]}/${result.rows[0].filename}`
+      console.log(downloadFile, ' successfully downloaded')
+      res.download(downloadFile)
     }
   })
 })
 
 app.post('/api/delete', (req, res) => {
   console.log('Recieved delete request')
-  db.pool.query(`DELETE FROM files WHERE hash = $1`, [req.body.hash], (err, result) => {
+  db.pool.query(`DELETE FROM files WHERE hash = $1 AND user_id = $2`, [req.body.hash, req.body.userID], (err, result) => {
     if (err) {
-      console.log('DB Failed', err)
+      console.log('Database deletion failed', err)
       res.send(err)
     } else {
-      console.log('DB Success')
-      fs.unlink(`./db/files/${req.body.filename}`, (deleteErr, deleteResult) => {
+      fs.unlink(`./db/files/${req.body.userID}/${req.body.filename}`, (deleteErr, deleteResult) => {
         if (deleteErr) {
-          console.log('Delete Failed', deleteErr)
+          console.log('Delete failed', deleteErr)
           res.send(deleteErr)
         } else {
-          console.log('Delete Success')
+          console.log(`${req.body.filename} successfully deleted by user ${req.body.userID}`)
           res.send(deleteResult)
         }
       })
@@ -74,6 +78,7 @@ app.post('/api/register', (req, res) => {
       console.log(err)
       res.status(500).send()
     } else {
+      console.log(`${req.body.username} has registered`)
       fs.mkdirSync(path.join(__dirname, "../", "db", "files", JSON.stringify(results.rows[0].id)))
       res.send(results.rows)
     }
@@ -106,7 +111,14 @@ app.post('/api/fileUpload', (req, res) => {
       var hashed = hash(file.name)
       file.mv('./db/files/' + req.body.userID + '/' + file.name, () => {
         db.pool.query(`INSERT INTO files(user_id, filename, hash, url, date_created, size)
-        VALUES ($1, $2, $3, $4, current_timestamp, $5)`, [req.body.userID, file.name, hashed, `${serverAddress}/api/download/${req.body.userID}/${hashed}`, file.size])
+        VALUES ($1, $2, $3, $4, current_timestamp, $5)`, [req.body.userID, file.name, hashed, `${serverAddress}/api/download/${req.body.userID}/${hashed}`, file.size], (err, result) => {
+          if (err) {
+            res.status(500).send()
+          } else {
+            console.log(`Uploaded ${file.name} from user ${req.body.userID}`)
+            res.send('success')
+          }
+        })
       })
     }
   } catch (err) {
